+++ collect_app/src/androidTest/java/org/odk/collect/android/utilities/FormIndexSavepointTest.java
@@ -49,33 +49,30 @@ public class FormIndexSavepointTest {
     @Test
     public void testBeginningOfForm() {
         FormIndex originalFormIndex = FormIndex.createBeginningOfFormIndex();
-        File tempIndex = SaveToDiskTask.savepointIndexFile(instancePath);
-        SaveToDiskTask.exportFormIndexToFile(originalFormIndex, tempIndex);
+        File tempIndex = SaveToDiskTask.savepointFormIndexIndexFile(instancePath);
+        FileUtils.exportFormIndexToFile(originalFormIndex, tempIndex);
 
-        FormIndex readFormIndex = FileUtils.loadFormIndex();
-
+        FormIndex readFormIndex = FileUtils.loadFormIndexFromFile();
         assertFormIndex(originalFormIndex, readFormIndex);
     }
 
     @Test
     public void testEndOfForm() {
         FormIndex originalFormIndex = FormIndex.createEndOfFormIndex();
-        File tempIndex = SaveToDiskTask.savepointIndexFile(instancePath);
-        SaveToDiskTask.exportFormIndexToFile(originalFormIndex, tempIndex);
+        File tempIndex = SaveToDiskTask.savepointFormIndexIndexFile(instancePath);
+        FileUtils.exportFormIndexToFile(originalFormIndex, tempIndex);
 
-        FormIndex readFormIndex = FileUtils.loadFormIndex();
-
+        FormIndex readFormIndex = FileUtils.loadFormIndexFromFile();
         assertFormIndex(originalFormIndex, readFormIndex);
     }
 
     @Test
     public void testNullReference() {
         FormIndex originalFormIndex = new FormIndex(1, 2, null);
-        File tempIndex = SaveToDiskTask.savepointIndexFile(instancePath);
-        SaveToDiskTask.exportFormIndexToFile(originalFormIndex, tempIndex);
+        File tempIndex = SaveToDiskTask.savepointFormIndexIndexFile(instancePath);
+        FileUtils.exportFormIndexToFile(originalFormIndex, tempIndex);
 
-        FormIndex readFormIndex = FileUtils.loadFormIndex();
-
+        FormIndex readFormIndex = FileUtils.loadFormIndexFromFile();
         assertFormIndex(originalFormIndex, readFormIndex);
     }
 
@@ -83,11 +80,10 @@ public class FormIndexSavepointTest {
     public void testNonNullReference() {
         TreeReference treeReference = TreeReference.rootRef();
         FormIndex originalFormIndex = new FormIndex(1, 2, treeReference);
-        File tempIndex = SaveToDiskTask.savepointIndexFile(instancePath);
-        SaveToDiskTask.exportFormIndexToFile(originalFormIndex, tempIndex);
+        File tempIndex = SaveToDiskTask.savepointFormIndexIndexFile(instancePath);
+        FileUtils.exportFormIndexToFile(originalFormIndex, tempIndex);
 
-        FormIndex readFormIndex = FileUtils.loadFormIndex();
-
+        FormIndex readFormIndex = FileUtils.loadFormIndexFromFile();
         assertFormIndex(originalFormIndex, readFormIndex);
     }
